Working Button for Counting Calories
added working button and selection features for the list of foods/items. Also a parser that lets you get the total amount of calories selected. Currently that number is displayed as a toast after selecting the confirm button, eventually we can move that over to displaying on home screen and tracking that way.